Add game parameters, modify order times for different houses and implement penalty and time-dependent satisfaction bar (#32)
* Add basic protocols for game logic

* Add basic implementation for GameControllerManager

* Implement basic item and order generation, satisfaction bar

* Remove xcuserdata files

* Remove more xcuserdata files

* Remove all xcuserdata files

* Remove xcuserdata files

* Add basic protocols for game logic

* Add basic implementation for GameControllerManager

* Implement basic item and order generation, satisfaction bar

* Remove xcuserdata files

* Remove more xcuserdata files

* Remove all xcuserdata files

* Resolve conflicts

* Modify GameController to include model properties, modify notifications, and minor additions to game logic

* Remove userdata files

* Fix code from errors in previous merge

* Add notifications for order changes, implement regular starting of orders

* Modify project.pbxproj file

* Modify project.pbxproj file

* Reorganise notifications, add resume round, get items from current active package, add new package methods

* Add timer to terminate game if game is paused for 30 seconds

* Reset satisfaction bar at start of each round

* Fix satisfaction bar bug where wrong package increases satisfaction

* Fix order time out and satisfaction bar bugs

* Extract host methods into subclass

* Add new RhythmicItem

* Add GameParameters and modify item and order allocation and generation

* Modify order times based on house satisfaction factor and implement penalty when satisfaction level reaches 0

* Implement time dependent satisfaction bar

+++ Relentless/Relentless/Network/Network.swift
@@ -41,7 +41,7 @@ protocol Network {
     func startRound(gameId: Int, roundNumber: Int)
     
     /// This is called by the host player to terminate the current round.
-    func terminateRound(gameId: Int, roundNumber: Int, satisfactionLevel: Int)
+    func terminateRound(gameId: Int, roundNumber: Int, satisfactionLevel: Float)
 
     /// This is called to pause the current round.
     func pauseRound(gameId: Int, currentRound: Int)
